requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
dateparser==0.7.1
Flask==1.0.2
grpcio==1.19.0
idna==2.8
inflection==0.3.1
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-spec-validator==0.2.6
pathlib==1.0.1
pylint==2.3.1
python-consul==1.1.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
regex==2019.4.14
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.15.2
wrapt==1.11.1